+++ xesmf/tests/test_util.py
@@ -1,6 +1,9 @@
 import numpy as np
 import xesmf as xe
+import xarray as xr
 import pytest
+
+from . test_frontend import ds_in
 
 
 def test_grid_global():
@@ -35,3 +38,42 @@ def test_cf_bnds_2d():
     eb = xe.util.cf_to_esm_bnds_2d(ds['lat_bnds'])
     np.testing.assert_array_equal(eb, ds['lat_b'])
 
+
+def test_cf_lon_lat():
+    # Dataset
+    lon, lat = xe.util.cf_lon_lat(ds_in)
+    xr.testing.assert_equal(lon, ds_in["lon"])
+    xr.testing.assert_equal(lat, ds_in["lat"])
+
+    # DataArray
+    lon, lat = xe.util.cf_lon_lat(ds_in["data"])
+    xr.testing.assert_equal(lon, ds_in["lon"])
+    xr.testing.assert_equal(lat, ds_in["lat"])
+
+    # Dict
+    lon, lat = xe.util.cf_lon_lat(dict(ds_in["data"].coords))
+    xr.testing.assert_equal(lon, ds_in["lon"])
+    xr.testing.assert_equal(lat, ds_in["lat"])
+
+    # Missing attributes
+    ds_miss = ds_in.copy()
+    ds_miss.lon.attrs["long_name"] = ""
+    with pytest.raises(ValueError):
+        xe.util.cf_lon_lat(ds_miss)
+
+    ds_miss = ds_in.copy()
+    ds_miss.lat.attrs["long_name"] = ""
+    with pytest.raises(ValueError):
+        xe.util.cf_lon_lat(ds_miss)
+
+
+def test_cf_bnds():
+    ds_miss = ds_in.copy()
+    ds_miss.lat.attrs.pop("bounds")
+    with pytest.raises(ValueError):
+        xe.util.cf_bnds(ds_miss, ds_miss["lat"])
+
+    ds_miss = ds_in.copy()
+    ds_miss.lat.attrs["bounds"] = "kabong"
+    with pytest.raises(ValueError):
+        xe.util.cf_bnds(ds_miss, ds_miss["lat"])
